+++ crates/translator/src/bin/translator-bin/cli.rs
@@ -28,6 +28,7 @@ impl Cli {
                          .about("translate all input from stdin"))
             .arg(arg!(-d --debug ...  "Turns on debugging and logging to file"))
             .arg(arg!(-j --json ... "Only generate dot files in JSON format"))
+            .arg(arg!(-p --pretty ... "Pretty print output"))
             .arg(arg!(-o --outputFolderSuffix <OUTPUT_FOLDER_SUFFIX> "The output folder suffix"));
 
         Self { cmd }