+++ crates/translator/src/java/mod.rs
@@ -10,7 +10,8 @@ pub extern "system" fn Java_be_ugent_idlab_knows_mappingLoom_Translator_translat
 ) -> JString<'local> {
     let outcome = unowned_env.with_env(|env| -> Result<_, jni::errors::Error> {
         let input_str = input.to_string();
-        let translated_res = process_one_str(&input_str);
+        let translated_res = process_one_str(&input_str)
+            .unwrap_or_else(|| "Error while translating the mapping".to_string());
         JString::from_str(env, translated_res.as_str())
     });
     outcome.resolve::<jni::errors::ThrowRuntimeExAndDefault>()