+++ crates/translator_rml/src/parser/extractors/io.rs
@@ -2,9 +2,7 @@
 //! Contains functionalities to deal with IO operations for parsing/translating 
 //! [RML v1.1.2 document](https://rml.io/specs/rml/)
 //!
-use std::fs::File;
-use std::io::{BufRead, BufReader, Read};
-use std::path::PathBuf;
+use std::io::BufRead;
 
 use sophia_api::source::TripleSource;
 use sophia_inmem::graph::FastGraph;
@@ -88,57 +86,17 @@ pub fn parse_str(input_str: &str) -> ExtractorResult<Document> {
     try_create_document(triples_maps, base_iri)
 }
 
-/// Parses the given file of an RML document into the 
-/// data model [Document].
-///
-/// # Errors
-///
-/// Returns an error for the following cases: 
-/// 1) Data model parsing error
-/// 2) File extension is not ".ttl"
-pub fn parse_file(path: PathBuf) -> ExtractorResult<Document> {
-    if let Some(ext) = path.extension() {
-        if ext != "ttl" {
-            return Err(ParseError::ExtensionError(format!(
-                "Extension does not exist {}",
-                ext.to_str().unwrap()
-            ))
-            .into());
-        }
-
-
-        let mut input_string = String::default();
-        BufReader::new(File::open(path.clone())?).read_to_string(&mut input_string)?;
-        
-        // Add default @base if none present
-        if !input_string.contains("@base") {
-            input_string = format!("@base <http://example.com/base/> .\n{}", input_string);
-        }
-        
-        let graph = load_graph_str(&input_string)?;
-        let triples_maps = extract_triples_maps(&graph)?;
-        let base_iri = input_string.split('\n').filter_map(extract_base_iri).next();
-
-        println!("base_iri: {:?}", base_iri);
-        return try_create_document(triples_maps, base_iri);
-    }
-
-    Err(ParseError::IOErrorStr(format!(
-        "File can't be read {}",
-        path.to_str().unwrap()
-    ))
-    .into())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::test_case;
+    use std::fs;
+    use std::path::PathBuf;
 
     #[test]
     fn one_tm_test() -> ExtractorResult<()> {
         let path = PathBuf::from(test_case!("rml/sample_mapping.ttl"));
-        let parsed_res = parse_file(path)?;
+        let parsed_res = parse_str(&fs::read_to_string(path)?)?;
 
         // One TriplesMap should be parsed
         assert_eq!(parsed_res.triples_maps.len(), 1);
@@ -149,7 +107,7 @@ mod tests {
     #[test]
     fn multiple_tm_test() {
         let path = PathBuf::from(test_case!("rml/multiple_tm.ttl"));
-        let parsed_res = parse_file(path);
+        let parsed_res = parse_str(&fs::read_to_string(path).unwrap());
 
         assert!(parsed_res.is_ok());
         // One TriplesMap should be parsed